refactor nn
