+++ notebooks/nn_init.py
@@ -21,7 +21,7 @@ from torch import nn
 
 # %%
 from aioway.fake import NnInitFn
-from aioway.hop import find_nn_init
+from aioway.nn import find_nn_init
 
 # %%
 nn_init = find_nn_init(NnInitFn(func=nn.Linear, in_features=3, out_features=5))
