+++ src/aioway/compilers/emits.py
@@ -7,7 +7,8 @@ from collections import abc as cabc
 
 from aioway._utils import decomp_flatten
 from aioway.builders import TensorBuilder
-from aioway.hop import Hop, Linear, ListHop, TensorHop
+from aioway.hop import Hop, ListHop, TensorHop
+from aioway.nn import Linear
 from aioway.spaces import ShapeSpace, Space
 
 __all__ = ["Emitter", "emitter_dcls", "JustLinearEmitter"]
