+++ src/aioway/hop/__init__.py
@@ -3,6 +3,8 @@
 from .funcs import *
 from .hop import *
 from .iters import *
+from .joins import *
 from .loaders import *
+from .maps import *
 from .nn import *
-from .relalg import *
+from .views import *
