--- src/aioway/hop/relalg/__init__.py
@@ -1,7 +1,0 @@
-# Copyright (c) AIoWay Authors - All Rights Reserved
-
-"Relation algebra related stream utilities."
-
-from .joins import *
-from .maps import *
-from .views import *
