+++ src/aioway/nn/containers.py
@@ -5,8 +5,8 @@ import typing
 from torch import nn
 
 from aioway.fake import NnInitFn
+from aioway.hop import NnLayerHop
 
-from .hop import NnLayerHop
 from .modules import NnInit, nn_init_dcls
 
 __all__ = ["Sequential"]
