+++ src/aioway/nn/losses.py
@@ -4,7 +4,8 @@ import typing
 
 from torch import nn
 
-from .hop import NnLossHop
+from aioway.hop import NnLossHop
+
 from .modules import NnInit, nn_init_dcls
 
 __all__ = [
