+++ src/aioway/nn/norms.py
@@ -4,7 +4,8 @@ import typing
 
 from torch import nn
 
-from .hop import NnLayerHop
+from aioway.hop import NnLayerHop
+
 from .modules import NnInit, nn_init_dcls
 
 __all__ = [
