+++ tests/compilers/test_emitters.py
@@ -8,7 +8,8 @@ from torch import nn
 from aioway.algos import SupervisedAlgo
 from aioway.compilers import JustLinearEmitter
 from aioway.dsets import TensorListHop, TensorStream
-from aioway.hop import Linear, ListHop, LoaderOpt, NnHop, NnInit, NnLayerHop, TensorHop
+from aioway.hop import ListHop, LoaderOpt, NnHop, NnLayerHop, TensorHop
+from aioway.nn import Linear, NnInit
 from aioway.spaces import Attr, AttrSpace, Shape, ShapeSpace
