+++ tests/hop/nn/test_hop.py
@@ -9,15 +9,8 @@ from torch import nn
 from aioway._utils import AnyDict, is_fake_tensor, torch_fake_mode
 from aioway.dsets import TensorListHop
 from aioway.fake import NnInitFn
-from aioway.hop import (
-    Hop,
-    NnLayerHop,
-    NnLossHop,
-    StackHop,
-    TensorHop,
-    build_nn_hop,
-    hop_cache_on,
-)
+from aioway.hop import Hop, NnLayerHop, NnLossHop, StackHop, TensorHop, hop_cache_on
+from aioway.nn import build_nn_hop
 
 
 @pytest.fixture
